Move duplicate service to controller and rename service to services
